+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - Added `JavaObjects.notEquals` which returns true if the given objects are not equal (opposite of `Objects.equals`).
 - Added `LeakDetectionLevel.modifyGlobal` method to change the global leak detection level.
 - Added `ExecutionWrappers.time` with configurable logging level. 
+- Added `LoggerAdapter.LoggingLevel.toJulLevel` to convert a `LoggingLevel` to JUL `Level`.
+- Moved `JulLoggerAdapter` class from tests to `*.lang.logging` package for easy logging without bothering with logging frameworks.
+- Moved `OneLineFormatter` class from tests to `*.lang.logging` package to configure JUL logging to be similar to SLF4J.
 
 ---